fix(network): move mynard2010 junction-loss model into combaero package
MPCEv2Element/ConstantKTeeElement import junction_loss_coefficient from
validation.junction.models.mynard2010 -- a dev-only top-level directory
that is excluded from the built wheel. A real `pip install combaero`
(or combaero-gui, which depends on it) raised ModuleNotFoundError: No
module named 'validation' the moment either class was constructed.
First caught by v0.4.0's combaero-gui publish verify job, which installs
the built wheel into a clean venv rather than running from a repo
checkout, so the module was never on sys.path.

Moves the self-contained implementation (only depends on numpy/math/
dataclasses) into python/combaero/network/_mynard2010.py, alongside the
existing _mpce_v2_jacobian internal module. validation/junction/models/
mynard2010.py becomes a thin re-export so existing validation scripts
and tests keep their import path unchanged.

Verified: built a fresh wheel and imported
combaero.network.mpce_v2_element from /tmp with no repo checkout on
the path -- previously failed, now clean. Full python/tests/ suite
(1579 passed) and check-python-style.sh also green.

Cuts CHANGELOG [Unreleased] (this fix + the publish-gui.yml race fix
from #233) into [0.4.1].
